SPA: Tracking the instruction text base pointer
Tracking the instruction text base pointer is hard, but it needs to
be done in order to ensure that GPU virtual memory mapping occurs
correctly. This patch is a hack which may work in most cases, but
in order to be completely correct, the text base has to be tracked
for each PTX file segment that is read in.