Shader Core: Fix failed assertion from changeset 68:e503e68d17dd
Didn't qfinish all patches before checking in changeset 68:e503e68d17dd, so this
assert was not removed appropriately. This fixes the failed assertion that will
occur on any write through the shader core, since the entries of the read_buffer
are not erased as they are from the iterator.