SPA: Make responsible for stream completion
Previously, the copy engine was responsible for signalling completion of
stream memory operations. However, viewing the SPA as the hardware side of
the hardware-software interface suggests that the copy engine should only
be considered a component of the SPA and thus carry no responsibility
besides signalling the SPA when completing a hardware memory operation.

In addition to the SPA signalling stream completion, the new
finishCopyOperation function also takes responsibility for reactivating
the CPU thread context after a copy operation completes, responsibility
that was previously held by the copy engine.