Thread Contexts: Make dynamic throughout sim
Tie the thread context to stream operations. When a kernel or memory operation
is launched from gpu_syscalls.cc, pass the thread context with the operation
into the stream processor array. Within the stream processor array, the thread
context is now only live during the life of the stream operation, so that
different thread contexts can access the GPU in interleaved ways.