GPUOptions: Parameterize CTAs per SM
Add a parameter to the simulation to specify the number of CTAs that can be
scheduled to an SM.

NOTE: Since the number of CTAs per SM is a free variable until bound when
GPGPU-Sim is initialized and this variable is not fixed anywhere in
gem5-fusion, it can be changed across checkpoint restore runs. This
flexibility is nice for testing the effects of variable amounts of concurrency
on the memory system.