CUDA: Add a thread blocking GPU syscall
cudaBlockThread is called from libcuda using m5_gpu(83), and the function
passes a pointer to a cache-line sized and aligned memory address "is_free"
in the simulated system that should be updated to True when the thread context
is free to continue executing. The thread context can be suspended in special
cases where the StreamProcessorArray is still executing memcpys or kernels.