GPGPUSim_Init: Refactor GPGPU-Sim Initialization
GPGPU-Sim used a lazy call to GPGPUSim_Init to initialize the GPU simulator,
and in gem5-gpu, we can remove that simply by initializing GPGPU-Sim when we
instantiate a gem5 GPU object. This patch moves all initialization into the
StreamProcessorArray constructor, removes unnecessary code, enables multiple
CUDA GPUs in the CUDA API, and tracks the device properties in the
StreamProcessorArray.