GPU Protocols: Set controller parents to ruby_system
This fixes a bug introduced in gem5 in which the source and destination of a
message buffer must be clocked at the same rate due to timing checking in the
destination that ensures time correctness. Since the parent of all the internal
Ruby components is the Ruby system itself (from which internal components get
their clock), also make the Ruby system the parent of the cache controllers.