ShaderLSQ: Temporarily set coalesce latency to 1
Currently, the way that the coalesceEvent is scheduled causes the coalescer to
be serialized instead of pipelined when multiple warps need coalescing, so the
throughput of the LSQ is limited to 1 request per coalescingLatency cycles
(e.g. 1 in every 4 cycles prior to this patch). This causes poor bandwidth to
the L1 caches when the latency is set to anything greater than 1, and currently
causes slowdowns of up to 4x for benchmarks with high occupancy and memory
demand.

This patch acts as a stopgap to get simulated system performance back to
something comparable to the simulator prior to introduction of the ShaderLSQ.
To fix this issue correctly will require splitting the processCoalesceEvent
function into 2 separate event-driven functions, 1 to schedule the coalescing
and 1 which steps the request through the rest of the LSQ pipeline to decouple
the coalescingLatency from insertion/removal from the outgoingBuffer.
The outgoing buffer will need to be removed in favor of memory request
injection into the memory system, and since the caches are clocked faster than
the LSQ, yet another event will need to be added to schedule the LSQ cycle
following the return of the data to process it and if necessary, writeback the
received data. Given the magnitude of these required changes, this stopgap is a
more immediate solution to the simulated system performance issue. There is
still a bug in serialization of requests on a per-cache-line basis, which is
causing up to 5x slowdowns for benchmarks with high occupancy, memory demand
and locality.

NOTE: In order to accurately model Fermi and provide flexibility in the
modeling of the LSQ, memory request scheduling should be designed to block on
threading granularity other than per-warp (e.g. half-warp), and requests from a
single blocking unit should not proceed in subsequent cycles, but instead wait
until the prior request from the blocking unit is complete. This will allow all
warps to proceed in lockstep into and out of the LSQ, something which is not
currently supported (even in GPGPU-Sim).

A redesigned ShaderLSQ is in the works and is forthcoming.