CEExitCallback: Make a member of GPUCopyEngine
The exit callback for the copy engine leaks when instantiated in the copy
engine constructor, because there isn't a way to delete it. Since there
should only be one exit callback per copy engine, make CEExitCallback a
member of GPUCopyEngine and instantiate it as a variable instead of as a
pointer.