ShaderLSQ: Forward flushall message to Ruby.
This patch changes the LSQ to forward the flush message from the cuda core to
Ruby when required.
This is necessary for coherence protocols which are not fully coherent to be
correct across kernel invocations and memory fences. Fences are not implemented
yet.
Currently forward flush is a parameter of the LSQ and set to true if using
VI_hammer and the command line option flush_kernel_end is provided.
Later we should change the default to be flushing at kernel end as it is
specified in the NVIDIA docs.