ShaderLSQ: Ground-up Redesign
The prior ShaderLSQ was buggy in it's handling of coalescer pipelining and
serializing memory requests into the memory system. This patch redesigns the
LSQ to more clearly and flexibly pipeline the stages and issue requests in
back-to-back cycles. It includes a pool of warp instruction buffers that allows
multiple requests from a single warp to be in flight, and cache accesses are
pushed to the cache as eagerly as possible without reordering. The net of these
enhancements allows up to 4x improved request bandwidth from a single GPU core.