Configs: Appropriately connect GPU, MMU, and CE ports
The previous code for connecting GPU, MMU and CE ports to the Ruby sequencers
caused aliasing for some sequencers and missing connections for others. This
patch fixes the connections and adds notes to clarify how these ports should
be connected. This also adds a fatal in the case that the user tries to use
a split address space, but still access the host's pagetable from GPU/CE.