CudaCore/ShaderLSQ: Separate out a control port
Add a separate port for the CudaCore to send control messages to the ShaderLSQ.
This change eliminates the overloaded use of the ShaderLSQ lane port 0 for
flush messages, which reduces potential complexity for multiplexing the port.
Further patches will use this port for propagating control signals between the
GPGPU-Sim shader issue stage and the LSQ for correct GPU fences.