ShaderMMU: Protect Against SE Mode Translation Faults
Previously, it was believed that faults could not be raised during page walks,
because the page table dynamically adds mappings as necessary in SE mode.
Unfortunately, this is only true for pages that are not protected. For
instance, NULL pointer dereferences still raise page faults, which cannot be
handed to the CPU for OS resolution in SE mode. Add a guard against calling the
ShaderMMU page fault handler in SE mode.