configs: Updates for GPGPU-Sim v3.2.2
Merge the configuration file changes from the latest GPGPU-Sim version into the
gpgpusim.template.config. This version of GPGPU-Sim now allows for setting core
execution unit timing parameters and finer-grained controls on caches and
memory.

With the move to intersim2, GPGPU-Sim is very temperamental about ensuring that
the actual node counts match the interconnect dimensions. To ensure this,
parameterize the new interconnect config file as a template, and calculate the
correct number of nodes in the config script. Fill in the template and place in
the output directory as with gpgpusim.config.