Atomics: Implement in CudaCore, ShaderLSQ, WarpInstBuffer
All the code required to accept atomic instructions from GPGPU-Sim, convert
the parameters and data in CudaCore, and send and receive memory accesses
between the ShaderLSQ and Ruby cache hierarchy. Atomic operation requests (on
a per-lane basis) are packaged into a coalesced access, which Ruby handles.

From testing, we have established that Fermi sends up to 3 atomic requests to
sublines (32B) of the L2 cache, and accesses to separate sublines of the same
line can proceed in parallel. Based on this, package an appropriate number of
atomic requests into CoalescedAccess to get appropriate bandwidth and latency.

NOTES:
1) This patch requires modification to the VI_hammer protocol for correct
atomic access handling and timing. It also requires a small patch on gem5 to
allow Ruby to complete the memory accesses in the hit callback.