Cuda GPU: Refine stats dump process
Using the Pseudoinst version of dumpresetstats schedules an event for 0 ticks
in the future and the event is executed after the rest of the code controlling
the GPU kernel start/end. This disallows stats to be accurately updated during
the functions that call dumpresetstats. Instead, directly call the Stats
functions dump() and reset() so that stats updates made in beginRunning and
finishKernel can be appropriately timed before or after the stats dump.

NOTE: This change causes includes to be sorted to pass style