Util: Add a cuobjdump utility
Debugging GPU application memory bugs is very difficult in the simulator. Some
things have been added to improve debugging, including DPRINTF functionality in
ShaderLSQ, and printing the program counter on GPU page faults. This utility
uses cuobjdump to print PTX code from a benchmark, and each line is prefixed
with the PC of the instruction. For consistency, the PC defaults to printing in
hex, but the --decimal option prints decimal PCs.