ShaderMMU: Add comments about setMiscRegNoEffect calls
The gem5-gpu GPU MMU updates CPU miscellaneous registers as a way to raise GPU
page faults to the CPU core. This is a convenient hack around actually
implementing a memory-mapped interrupt from the GPU. This process is invasive,
not an accepted practice in gem5, and may cause different side-effects when
simulating different gem5 core types.

In the O3 CPU, specifically, setting registers outside the context of executing
instructions causes the CPU to squash the pipeline and reexecute in-flight
instructions in order to ensure the core reads the correct values from the
modified registers. With gem5-gpu, this squashing causes CUDA blockThread()
calls to reexecute the thread context suspend, frequently causing the thread
context to remain suspended for excessively long.

Add comments to make it clear that setMiscRegNoEffect calls are potentially
buggy with some core types.