CudaGPU: Update gpgpu_sim initialization
The prior initialization of GPGPU-Sim was structured using left-over code from
when GPGPU-Sim was instantiated from the first CUDA call. Since we are now
instantiating GPGPU-Sim in a well-defined way, there is no need to separate the
calls for passing the CudaGPU to the gpgpu_sim instance. Unify these calls.