CudaGPU: Fix thread blocking before GPU fault
There is a possible race condition that exists with CUDA library thread
suspension: If a GPU fault is raised between when a CPU thread finishes a prior
fault and before it calls blockThread() again, the CPU thread may get suspended
before the fault is serviced. This can result in blocked GPU progress until the
CPU thread is reactivated by another interrupt (e.g. timer interrupt), leading
to exceptionally low GPU IPC. To fix this, just check in blockThread() whether
there is an outstanding fault, and if so, do not suspend the thread. Allow it
to continue executing to finish the fault and then try suspending again.