ShaderMMU: Set miscellaneous register with actually no effect
O3 CPU pipeline squashes cause numerous problems with GPU page faults. A
specific example is when the CPU thread calls cudaBlockThread to suspend the
running thread, but the GPU raises a pagefault before the CPU is suspended. The
pipeline squash caused by setting CPU registers from the GPU can cause the
suspend pseudo-instruction to be squashed, but the GPU tries to wake up the
thread which is not suspended.

To avoid this issue use a new variant of the setMiscRegNoEffect() function that
disallows pipeline squashing when it is known that the CPU need not drain. This
function will also be used in other forthcoming fixes to GPU page fault
handling.