cpu: Compile and allow to run MinorCPU
It appears that the MinorCPU can build with both x86 and ARM, and short tests
indicate that it also works to run SE mode simulations for both. Compile the
core by default, and adjust config scripts to allow MinorCPU sims to be run.

Note: We might want to add O3CPU and MinorCPU regression tests in the future.