ShaderTLB, ShaderMMU: Refactor virtual and physical page base addresses
The variables vpn and ppn were meant to refer to virtual and physical page
number, respectively, but these names are misleading. Page numbers are not
full addresses, but rather the top bits of the address with the page offset
removed. Throughout gem5-gpu, address translations work with virtual and
physical page base addresses (i.e. full width of the address). gem5 uses the
variable names vaddr and paddr, which fail to reflect that these are base
addresses for the pages. To avoid confusion, rename vpn -> vpBase/vp_base, and
ppn -> ppBase/pp_base.