ShaderMMU: Refactor finalizeTranslation
The looping and loop conditionals in finalizeTranslation were confusing and
difficult to correctly modify without introducing memory bugs/leaks. In fact,
the current code allows the passed translation to be deleted before all other
translations can be completed, possibly resulting in a segmentation fault.
Refactor the code to be clearer and fix the memory bug.