Memsets: Shift some to copy engine
Memsets are required to be performed by the CPU when the copy engine or GPU
does not have a way to translate addresses. This only occurs when the GPU is
not managing it's own page tables *and* the GPU does not have access to the
host's page table (i.e. with --access-host-pagetable). If allowed to access
the host's page table, move cudaMemsets back to the copy engine.