ShaderTLB: Fix insert entry initialization
In the TLBMemory::insert() function, when all entries in a way have the same
replacement time equal to the curTick(), the entry local variable would not get
initialized in the search for an entry to replace. This would cause the
following assertion to fail. Allow the search to pick an entry with mruTick
equal to curTick() to avoid the problem.