ShaderLSQ: Add a deadlock check
When running in FS mode, ShaderLSQ instructions can take thousands of cycles to
complete when waiting on the page fault handler. Add a deadlock check to ensure
the LSQ is making progress. 1,000,000 cycles was chosen to be significantly
longer than a pagefault (up to 15,000 cycles), but not unreasonably long to
wait for a deadlock before exiting.