CudaGPU, ShaderMMU: Preliminary thread handling capabilities
When running in full-system mode, the CPU process may get migrated among cores,
leading the GPU to raise page faults to an incorrect CPU core. As a first step
toward more robust CPU thread handling and GPU context management, add thread
context checks and allow the context to be updated if CUDA syscalls detect a
new thread context.

Further, this patch, overall, offers far more descriptive warnings. The only
way to detect that the OS is migrating a CPU thread is to check the thread
context pointer and/or CR3 when the thread makes a call into CUDA syscalls to
the ShaderMMU. Unfortunately, this can be a long time after a thread context
switch actually occurs, and the simulation could have fallen into some
incorrect state by that time. To start getting a handle on this issue, add a
warning when the CR3 register is incorrect when raising a page fault (i.e.
when the CR3 needs to be correct), so we can collect some examples where
simulation runs off the rails.

NOTE: More is needed to make this change really robust. Specifically, thread
migration typically happens during long periods of CPU inactivity, so it is
unlikely that one of the CUDA syscalls will be called after a thread context
migration, but before a translation is needed. This means it is unlikely that
the CUDA syscalls will adjust for a migrated thread without crashing yet. To
improve this change might be as easy as adding a GPU pseudoinstruction call
into thread migration or GPU pagefault interrupt microcode.